Allow for continual re-scanning and idempotently update devices in the repository
Wemo's are often fussy and don't always show up on your first scan so
allowing the thread to continually update or amend its own list will
resolve for those flakey UPNP wemo queries. This does however require
the user be aware that some wemos may actually go away and this will
cause some switches to be present but not functional in the Radar.